@@ -3,8 +3,6 @@ using System;
 using System.IO;
 using System.Windows.Controls;
 using NLog.Config;
-using NLog.LogReceiverService;
-using NLog.Targets;
 using Torch;
 using Torch.API;
 using Torch.API.Managers;
@@ -14,8 +12,6 @@ using Torch.Session;
 using S_Utilities.UI;
 using S_Utilities.Settings;
 using S_Utilities.Utils;
-using Torch.Patches;
-using Torch.Server.Views;
 
 // ReSharper disable ClassNeverInstantiated.Global
 
@@ -33,9 +29,13 @@ namespace S_Utilities
         private static Persistent<S_Config>? _config;
         public static S_Config? Config => _config?.Data;
 
+        public static bool IsOnline = false;
+        public static string InstName = "";
+
         public override void Init(ITorchBase torch)
         {
             base.Init(torch);
+            InstName = Torch.Config.InstanceName;
 
             SetupConfig();
 
@@ -47,19 +47,14 @@ namespace S_Utilities
 
             // Torch's new event handler logger things is still WIP so lets not get to cozy with it yet...
             LoggingConfiguration config = LogManager.Configuration ?? new LoggingConfiguration();
-            SenXCustomTarget customTarget = new SenXCustomTarget();
+            SenXCustomTarget customTarget = new ();
             config.AddTarget("SenXCustomTarget", customTarget);
-            LoggingRule rule = new LoggingRule("*", LogLevel.Trace, customTarget);
+            LoggingRule rule = new ("*", LogLevel.Trace, customTarget);
             config.LoggingRules.Add(rule);
             LogManager.Configuration = config;
             
-            SenXCustomTarget.CustomLogEventReceived += SenXLogEventReceived;
+            SenXCustomTarget.CustomLogEventReceived += Log2Discord_Processor.ProcessLogEvent;
             Save();
-        }
-
-        private static void SenXLogEventReceived(LogEventInfo obj)
-        {
-            
         }
 
         private void SessionChanged(ITorchSession session, TorchSessionState state)
@@ -68,17 +63,14 @@ namespace S_Utilities
             {
                 case TorchSessionState.Loaded:
                     Log.Info("Session Loaded!");
+                    IsOnline = true;
                     break;
 
                 case TorchSessionState.Unloading:
                     Log.Info("Session Unloading!");
+                    IsOnline = false;
                     break;
             }
-        }
-        
-        private void LogEventReceived(LogEvent logEvent)
-        {
-            Log.Info(logEvent.Message);
         }
 
         private void SetupConfig()
@@ -87,9 +79,7 @@ namespace S_Utilities
 
             try
             {
-
                 _config = Persistent<S_Config>.Load(configFile);
-
             }
             catch (Exception e)
             {
