@@ -6,7 +6,7 @@ namespace S_Utilities.Settings
 {
     public partial class S_Config
     {
-        private ObservableCollection<LogHandler> _LogHandlers = new ObservableCollection<LogHandler>();
+        private ObservableCollection<LogHandler> _LogHandlers = new ();
         public ObservableCollection<LogHandler> LogHandlers { get => _LogHandlers; set => SetValue(ref _LogHandlers, value); }
     }
 }
